fotos y comentarios de los usuarios

+++ app/Http/Controllers/PostController.php
@@ -6,7 +6,9 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
 use App\Models\Post;
+use App\Models\Comment;
 
 class PostController extends Controller
 {
@@ -23,44 +25,41 @@ class PostController extends Controller
         return view('posts_views.create', compact('user'));
     }
 
-    public function create(Request $request){
+    public function create(Request $request)
+    {
         $validator = Validator::make($request->all(), [
             'title' => 'required|max:255',
             'description' => 'required',
-            //'image' => 'image|mimes:jpeg,png,jpg,gif|max:2048',
-        ],[
-            'title.required' => 'El titulo is obligatorio.',
-            'title.max' => 'EL titulo debe tener menos de 255 caracteres.',
+            'image' => 'image|mimes:jpeg,png,jpg,gif|max:2048',
+        ], [
+            'title.required' => 'El título es obligatorio.',
+            'title.max' => 'El título debe tener menos de 255 caracteres.',
             'description.required' => 'La descripción es obligatoria.',
-            //'image.required' => 'La imagen es obligatoria.',
-            //'image.image' => 'El archivo debe ser una imagen.',
-            //'image.mimes' => 'La imagen debe ser de tipo jpeg, png, jpg o gif.',
-            //'image.max' => 'La imagen debe pesar menos de 2MB.',
-        ]
-        );
+            'image.image' => 'El archivo debe ser una imagen.',
+            'image.mimes' => 'La imagen debe ser de tipo jpeg, png, jpg o gif.',
+            'image.max' => 'La imagen debe pesar menos de 2MB.',
+        ]);
 
         if ($validator->fails()) {
             return redirect()->route('posts.form')
-                     ->withErrors($validator)
-                     ->withInput();
+                ->withErrors($validator)
+                ->withInput();
         }
-/*
+
+        $imagePath = null;
         if ($request->hasFile('image')) {
-            $imagePath = $request->file('image')->store('public/posts'); // Guarda en storage/app/public/posts
-            $imageName = str_replace('public/', '', $imagePath); // Para obtener la ruta relativa
-        } else {
-            $imageName = null;
+            $image = $request->file('image')->store('posts');
         }
-*/
+
         $post = new Post();
         $post->title = $request->title;
         $post->description = $request->description;
-        //$post->file_path = $imageName;
+        $post->image = $image;
         $post->user_id = Auth::id();
         $post->save();
 
         return redirect()->route('posts.home')
-                         ->with('success', 'Post creado correctamente.');
+            ->with('success', 'Post creado correctamente.');
     }
 
     public function like(Request $request){
@@ -83,11 +82,34 @@ class PostController extends Controller
     }
 
     public function delete($id)
-    {
-        $post = Post::findOrFail($id);
+    {   
+        $validator = Validator::make(
+            ['id' => $id],
+            ['id' => 'required|exists:App\Models\Post,id']
+        );
+
+        if ($validator->fails()) {
+            return redirect()->back()->withErrors($validator);
+        }
+
+        // Obtener el post y verificar si existe
+        $post = Post::find($id);
+        if (!$post) {
+            return redirect()->back()->with('error', 'El post no existe.');
+        }
+
+        // Eliminar comentarios asociados
+        Comment::where("post_id", $id)->delete();
+
+        // Eliminar imagen si existe
+        if ($post->image) {
+            Storage::delete('posts/' . $post->image);
+        }
+
+        // Eliminar el post
         $post->delete();
 
-        return redirect()->route('')
-                         ->with('success', 'Post eliminado correctamente.');
+        return redirect()->route('posts.home')
+            ->with('success', 'Post eliminado correctamente.');
     }
 }
